+++ client/src/components/LoginComponent.jsx
@@ -1,8 +1,56 @@
-import React from 'react';
+import React, { useState } from 'react';
+import '../index.scss';
 
-function LoginComponent({
-  handleCreateUser, handleLoginUser, clientDataObject, setClientData,
-}) {
+function LoginComponent({ setClientData }) {
+  // clientDataObject set to returned value of fetch
+  const handleCreateUser = async (createUserData) => {
+    try {
+      // make enclosing function async
+      const postURL = '/auth/create';
+      const fetchResponse = await fetch(postURL, {
+        method: 'POST',
+        headers: {
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(createUserData),
+      });
+      const data = await fetchResponse.json();
+      console.log(data);
+      return data;
+    } catch (error) {
+      // handle error
+      console.log(error);
+      return {};
+    }
+
+    // handleCreateDialogClose();
+  };
+  const handleLoginUser = async (loginUserData) => {
+    try {
+      // make enclosing function async
+      const postURL = '/auth/login';
+      const fetchResponse = await fetch(postURL, {
+        method: 'POST',
+        headers: {
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(loginUserData),
+      });
+      const data = await fetchResponse.json();
+      // figure out how to populate state from this response
+      console.log(data);
+      return data;
+
+      // setLayouts(data);
+    } catch (error) {
+      // handle error
+      console.log(error);
+      return {};
+    }
+  };
+
   const submitForm = (event) => {
     event.preventDefault();
     const buttonClicked = event.nativeEvent.submitter.id; // is there a better way to do this?
@@ -11,29 +59,28 @@ function LoginComponent({
     if (buttonClicked === 'loginButton') setClientData(handleLoginUser(userDataObject));
     else if (buttonClicked === 'createButton') setClientData(handleCreateUser(userDataObject));
   };
-  /* no way to gradually switch background colors */
-  // const switchToNight = (event) => {
-  //   console.log(document.getElementsByTagName('body')[0]);
-  //   document.getElementsByTagName('body')[0].style.background
-  //   = 'linear-gradient(#25003e 8%, #054a91 90%, #054A91 100%)';
-  // };
 
   return (
-    <form onSubmit={submitForm}>
-      <label htmlFor="username">
-        <p>Username:</p>
-        <input type="text" id="username" name="username" required />
-      </label>
-      <label htmlFor="password">
-        <p>Password:</p>
-        <input type="password" id="password" name="password" required />
-      </label>
-      {/* pass in login/sign up */}
-      <div className="loginButtonContainer">
-        <input type="submit" value="Login" className="submitButton" id="loginButton" />
-        <input type="submit" value="Create Account" className="submitButton" id="createButton" />
+    <>
+      <div className="loginContainer loginContainerBorder">
+        <form onSubmit={submitForm}>
+          <label htmlFor="username">
+            <p>Username:</p>
+            <input type="text" id="username" name="username" required />
+          </label>
+          <label htmlFor="password">
+            <p>Password:</p>
+            <input type="password" id="password" name="password" required />
+          </label>
+          {/* pass in login/sign up */}
+          <div className="loginButtonContainer">
+            <input type="submit" value="Login" className="submitButton" id="loginButton" />
+            <input type="submit" value="Create Account" className="submitButton" id="createButton" />
+          </div>
+        </form>
       </div>
-    </form>
+      <div className="waveBackground" />
+    </>
   );
 }
