+++ client/src/index.scss
@@ -25,8 +25,8 @@ body {
     height: 100%;
     width: 100%;
     display: flex;
-    // justify-content: center;
-    // align-items: center;
+    justify-content: center;
+    align-items: center;
 }
 
 .waveBackground {
