+++ package.json
@@ -17,7 +17,8 @@
     "express": "^4.18.2",
     "mongoose": "^7.2.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.2"
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
